+++ coolamqp/__init__.py
@@ -1,2 +1,2 @@
 # coding=UTF-8
-__version__ = '1.1.1_a2'
+__version__ = '1.1.1_a3'